add note with api
