+++ project_2/NotesApp/src/App.vue
@@ -1,170 +1,3 @@
 <template>
-  <main>
-    <div v-show="showModal" class="overlay">
-      <div class="modal">
-        <textarea
-          v-model="note"
-          name="note"
-          id="note"
-          cols="30"
-          rows="10"
-        ></textarea>
-        <p v-if="error">{{ error }}</p>
-        <button @click="addNote">Add Note</button>
-        <button class="close" @click="showModal = false">Close</button>
-      </div>
-    </div>
-    <div class="container">
-      <header>
-        <h1>Notes {{ showModal }}</h1>
-        <button @click="showModal = true">+</button>
-      </header>
-      <div class="cards-container">
-        <div
-          v-for="note in notes"
-          :key="note.id"
-          class="card"
-          :style="{ backgroundColor: note.backgroundColor }"
-        >
-          <p class="main-text">{{ note.text }}</p>
-          <p class="date">{{ note.date.toLocaleDateString("en-US") }}</p>
-          <button @click="delNote(note.id)">Delete</button>
-        </div>
-      </div>
-    </div>
-  </main>
+  <router-view></router-view>
 </template>
-
-<style lang="scss" scoped>
-main {
-  width: 100vw;
-  height: 100vh;
-  position: relative;
-  .overlay {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.77);
-    z-index: 10;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    .modal {
-      width: 750px;
-      background-color: #fff;
-      border-radius: 10px;
-      padding: 30px;
-      position: relative;
-      display: flex;
-      flex-direction: column;
-      button {
-        padding: 10px 20px;
-        font-size: 20px;
-        width: 100%;
-        background-color: blueviolet;
-        border: none;
-        color: #fff;
-        cursor: pointer;
-        margin-top: 10px;
-      }
-      .close {
-        background-color: rgb(193, 15, 15);
-      }
-      p {
-        color: rgb(193, 15, 15);
-      }
-    }
-  }
-
-  .container {
-    max-width: 1000px;
-    padding: 10px;
-    margin: 0 auto;
-
-    header {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-
-      h1 {
-        font-weight: bold;
-        margin-bottom: 25px;
-        font-size: 75px;
-      }
-
-      button {
-        border: none;
-        padding: 10px;
-        width: 50px;
-        height: 50px;
-        background-color: rgb(21, 20, 20);
-        border-radius: 100%;
-        color: #fff;
-        font-size: 20px;
-        cursor: pointer;
-      }
-    }
-
-    .cards-container {
-      display: flex;
-      flex-wrap: wrap;
-
-      .card {
-        width: 225px;
-        height: 225px;
-        background-color: rgb(237, 182, 44);
-        padding: 10px;
-        border-radius: 15px;
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-        margin-right: 20px;
-        margin-bottom: 20px;
-
-        .date {
-          font-size: 12.5px;
-          font-weight: bold;
-        }
-      }
-    }
-  }
-}
-</style>
-
-<script setup>
-import { ref } from "vue";
-
-const showModal = ref(false);
-const note = ref("");
-const notes = ref([]);
-const error = ref("");
-
-function getRandomColor() {
-  return "hsl(" + Math.random() * 360 + ", 100%, 75%)";
-}
-
-const addNote = () => {
-  if (note.value.length < 9) {
-    return error.value = "Note needs to be 10 characters or more"
-  }
-
-  notes.value.push({
-    id: Math.floor(Math.random() * 1000000),
-    text: note.value,
-    date: new Date(),
-    backgroundColor: getRandomColor(),
-  });
-
-  showModal.value = false;
-  note.value = "";
-  error.value = "";
-  console.log(notes.value);
-};
-
-const delNote = (noteId) => {
-  if (confirm("Are you sure you want to delete")) {
-    notes.value = notes.value.filter(note => note.id !== noteId)
-  }
-  console.log(notes)
-}
-</script>
