+++ project_2/NotesApp/src/main.js
@@ -2,5 +2,6 @@ import './assets/main.css'
 
 import { createApp } from 'vue'
 import App from './App.vue'
+import router from "./router";
 
-createApp(App).mount('#app')
+createApp(App).use(router).mount('#app')
